Add delete functionality to notes
Notes array have note objects as items.
Each note object has properties, id, title, content

delete function filters the note object items to check if the id of the note to be deleted is same as the iterated note bject's id.
if so that note object item is not returned.